Refactor level_count and protect with seqcount
We were manually manipulating the level counts in the super in a bunch
of places under the manifest rwsem.  This refactors them into simple get
and add functions.  We protect them with a seqcount so that we'll be
able read them without blocking (from trans hold attempts).  We also add
a helper for testing that a level is full because we already used
different comparisons in two call sites.

Signed-off-by: Zach Brown <zab@versity.com>