Don't prematurely write dirty super
A previous refactoring messed up and had scoutfs_trans_write_func()
always write the dirty super even when nothing was dirty and there was
nothing for the sync attempt to do.  This was very confusing and made it
look like the segment and treap writes were being lost when in fact it
was the super write that shouldn't have happened.

Signed-off-by: Zach Brown <zab@versity.com>