scoutfs: fix ring next/prev walk comparison test
The scoutfs_ring_next() and _prev() functions had a really dumb bug
where they check the sign of comparisons by comparing with 1.  For
example, next would miss that the walk traversed a lesser item
and wouldn't return the next item.

This was causing compaction to miss underlying segments, creating
segments in levels that had overlapping keys, which then totally
confused reading and kept it from finding the items it was looking for.

Signed-off-by: Zach Brown <zab@versity.com>