scoutfs: fix item read seg walk limit
When read items into the cache we have the range of keys that were
missing from the cache.  The item walk was stopping when it hit the end
of the missing cache range, not when it hit the end of the keys that
were covered by all the segments.

This would manifest as huge regions of missing items.  The read would
walk off the relatively closed end of the highest level segment.  It
would keep reading while there were items in the upper levels but all
those keys that would have been found in additional lower level segments
are missing.  Eventually it'd hit the end of the higher level sgement
and mark that region as cached.

With it fixed it now stops the read appropriately and will come around
next time to read the range that coveres the next lowest level segment.

Signed-off-by: Zach Brown <zab@versity.com>